uploader component refactored
